fixed props build issue
